move nsmb function into dedicated .py file
