+++ qetpy/core/__init__.py
@@ -2,3 +2,4 @@ from ._didv import *
 from ._iv import *
 from ._noise import *
 from ._fitting import *
+from ._of_nsmb import *
